keylogger/requirements.txt.txt
pywin32
pynput
scipy
cpytopgraphy
requests
pillow
sounddevice
